Update to work with dual mode in irene
Irene is able to write to separate output file if the DAQ is running
in dual trigger mode. To do this a new conf parameter is added to the
templates and a filename for the second file.

This commit adds a new flag for CERES (-dual). If active, will create
the directories pmaps/trigger1 and pmaps/trigger2 for those files.